// tests/utils/enums/AddressesEnum.ts
export enum AddressesEnum {
  WALLET_PUBLIC = '0xE181e40D3365769705f7c488FE81225237A25ad8',
  WETH_TOKEN = '0xc778417E063141139Fce010982780140Aa0cD5Ab',
  DXD_TOKEN_RINKEBY = '0x554898A0BF98aB0C03ff86C7DccBE29269cc4d29',
  DXD_TOKEN_MAINNET = '0xa1d65E8fB6e87b60FECCBc582F7f97804B725521',
  STRONG_TOKEN_RINKEBY = '0x990f341946A3fdB507aE7e52d17851B87168017c',
  XEENUS_TOKEN_RINKEBY = '0x022E292b44B5a146F2e8ee36Ff44D3dd863C915c',
  SECOND_TEST_WALLET = '0xaE579C6B177EeAF8A7B68653fDDC2A85AFaAD70f',
  DAI_TOKEN_RINKEBY = '0x5592EC0cfb4dbc12D3aB100b257153436a1f0FEa',
  USDC_TOKEN_RINKEBY = '0x4DBCdF9B62e891a7cec5A2568C3F4FAF9E8Abe2b',
  USDC_TOKEN_ARINKEBY = '0x1E77ad77925Ac0075CF61Fb76bA35D884985019d',
  LINK_ADDRESS_RINKEBY = '0x01BE23585060835E02B77ef475b0Cc51aA1e0709',
}